Random game selection and timer
Set random_start to true for a random game
Set starting game to the index you want if it's not random (starts on 0)

To set your game duration, add it to the GAME_TIMES array in the index.  Example: If you wanted game 3's duration to be 5, change the array to [4, 4, 5]. This is measured in seconds because why not.